Move azure/compute/vm/params.go to azure/compute/azureparams/params.go
